Include NHL team statistics
Team statistics for all NHL teams can be pulled with a single call to
the Teams class in sportsreference.nhl. The class returns a list of all
teams that participated in the requested season including every team's
season stats, such as the number of wins and losses, total goals for and
against, total points, penalty killing percentage, and more.

Signed-Off-By: Robert Clark <robdclark@outlook.com>