Include NBA team statistics
Team statistics for all NBA teams can be pulled with a single call to
the Teams class in sportsreference.nba. The class returns a list of all
teams that participated in the requested season including every team's
season stats, such as the number of wins and losses, total field goals
and free throws, number of blocks, offensive rebounds, as well as their
opponent's stats, such as total field goals against, and much more.

Signed-Off-By: Robert Clark <robdclark@outlook.com>