Include a utils file
The utils file handles common tasks for all of the various sports, such as
finding the starting month of the season and parsing the results of the HTML
contents. This reduces the amount of redundant code for the various classes
in every sport, allowing them to focus on the attributes and properties that
are specific to the class.

Signed-Off-By: Robert Clark <robdclark@outlook.com>