Include NFL team statistics
Team statistics for all NFL teams can be pulled with a single call to
the Teams class in sportsreference.nfl. The class returns a list of all
teams that participated in the requested season including every team's
season stats, such as the total number of yards and points, the average
yards per play, the number of passing touchdowns, and much more.

Signed-Off-By: Robert Clark <robdclark@outlook.com>