Include NCAAF team statistics
Team statistics for all NCAA Division-I football teams can be pulled with
a single call to the Teams class in sportsreference.ncaaf. The class
returns a list of all teams that participated in the requested season
including every team's season stats, such as the total number of yards
and points, the conference wins and losses, the number of rushing
touchdowns, and much more.

Signed-Off-By: Robert Clark <robdclark@outlook.com>