Include NCAAB team statistics
Team statistics for all NCAA Division-I Men's basketball teams can be
with a single call to the Teams class in sportsreference.ncaab. The
class returns a list of all teams that participated in the requested
season including every team's season stats, such as the total number of
conference wins and losses, number of steals, total assists, field goal
percentage, pace, and much more.

Signed-Off-By: Robert Clark <robdclark@outlook.com>