Include MLB team statistics
Team statistics for all MLB teams can be pulled with a single call to
the Teams class in sportsreference.mlb. The class returns a list of all
teams that participated in the requested season including every team's
season stats, such as the number of wins and losses, the total number of
home runs, times caught stealing, number of complete games, and much
more.

Signed-Off-By: Robert Clark <robdclark@outlook.com>