Include optional index for multiple fields
In some cases, multiple fields can refer to a different item but have
the same 'data-stat' tag (which is used for matching values with
attributes). For example, in baseball, the number of home runs a team
has hit is different from the number of home runs a pitcher has given
up. On sports-reference.com, however, both of these fields have the
'data-stat' tag of 'HR'. Returning the first element will return the
incorrect result in the second case. The index variable corresponds to
the order in which the fields appear in the html_data object. The
default action is to return the first value if no index is listed.

Signed-Off-By: Robert Clark <robdclark@outlook.com>