Reduce code in tests
The original implementation of the setup methods for the testing classes
was not supported by Python 3.3, so a redundant workaround was used
instead. Now, the setup_method can be used with different parameters to
allow code to be reused in all testing methods without using the same
code over and over.

Robert Clark <robdclark@outlook.com>