Add Python 3.7 support
Now that Python 3.7 has been officially released, it can be included
with the Travis-CI tests. Travis doesn't natively support version 3.7
due to the reliance on a version of OpenSSL that is not found in the
'trusted' repository, so 'xenial' is required as it contains all of the
proper code versions. Thanks to this repository update, the development
branches for both Python 3.7 and Python 3.8 can also be included in the
Travis-CI builds to ensure future releases for both versions are
supported upon release.

Signed-Off-By: Robert Clark <robdclark@outlook.com>