Genericize error parsing for pulling boxscores
Depending on the python version, various webpages that have invalid URLs
throw random errors which can't be safely caught using the HTTPError
from urllib. Unfortunately, the try...except block needs to be more
generic in order to handle the high number of possible errors that can
be thrown, which would be too complicated to elegantly and efficiently
handle.

Signed-Off-By: Robert Clark <robdclark@outlook.com>