Special parse goalie stats for NHL
NHL teams can use different goalies through the duration of the game.
As a result, not all of the information can be extracted from just
reading the first goalie's stats for each team. In addition, assuming
that the second goalie on the page plays for the home team is a false
assumption as the away team might have used multiple goalies.

Signed-Off-By: Robert Clark <robdclark@outlook.com>