Calculate 2-point field goals for NCAAB teams
The 2-point field goal attempts, baskets, and percentages are tracked by
the Boxscore class, but not the Team class. To enable better comparsions
between single-game performances and season statistics for a team,
especially in a machine-learning application, the two point field goals
should be calculated in the Team class by finding the difference between
total field goals and three-point field goals.

Signed-Off-By: Robert Clark <robdclark@outlook.com>