Calculate net efficiency rating for NCAAB teams
The net efficiency rating is equivalent to the offensive rating minus
the defensive rating and indicates if a team scores more or fewer points
than their opponents per 100 possessions. The Boxscore class has this
information, so it should be included with the Team class to allow
better integration between the datasets for machine-learning type
applications.

Signed-Off-By: Robert Clark <robdclark@outlook.com>