Create special version for developer builds
In anticipation of releasing developer builds to test.pypi on all stable
commits, a developer release number should be used to prevent re-use of
the main release tags which use semantic versioning. At the moment, the
dev release number is the most recent released version number with a
timestamp after the version.

Signed-Off-By: Robert Clark <robdclark@outlook.com>