Prevent errors on null boxscores
If a boxscore is requested from the Schedule class and returns a value
of None, it will throw an error while trying to build a DataFrame full
of None frames. All boxscores that don't exist should be skipped and, if
necessary, the DataFrame should be None if there are no valid frames.

Signed-Off-By: Robert Clark <robdclark@outlook.com>