Update Travis deployments to increment version
For all future patches (ie. any branch that begins with "patch/" must
increment the patch number in the semver version. For all other builds,
the minor number should be incremented by one to indicate that a build
is tracking towards a future release instead of a pre-release for an
already released package version.

Signed-Off-By: Robert Clark <robdclark@outlook.com>