Add specificity to error catching for requests
The requests were originally left blank but should instead contain
HTTP-specific error checking to be pycodestyle-compatible. Using the six
package as a wrapper, we don't have to worry about pulling urllib
libraries specific for each python major version and can use a single
import regardless of which python is running.

Signed-Off-By: Robert Clark <robdclark@outlook.com>