Allow NFL Boxscores to be iterable
Instead of requiring users to query boxscores over a single week,
multiple weeks should now be supported to allow users to query multiple
weeks worth of boxscores from a single call to the Boxscores class. Not
only will this be easier to use, but this method is also more performant
than looping over multiple calls to the Boxscores class.

Signed-Off-By: Robert Clark <robdclark@outlook.com>