Fix an issue pulling future boxscores
An error is currently thrown whenever the boxscore of a game that hasn't
completed (for example, any future games and games on the current day)
is requested from sportsreference as it tries to parse a score when none
exists. Instead of throwing an error, the score should just show as
None.

Signed-Off-By: Robert Clark <robdclark@outlook.com>