Create slim parameter for Roster class
A new slim parameter should be added to the Roster class if the desire
is to just get the names and/or player ID's of each player on a team.
This short-circuits the call to the Player class which eliminates the
need for a 2-3 second instantiation and call to the individual player's
stats page. This allows the names and IDs to be returned in a small
fraction of the time compared to the standard request.

Signed-Off-By: Robert Clark <robdclark@outlook.com>