Add individual player stats for NCAAF Boxscores
A major piece of information that was missing from the Boxscores class
was individual player stats in a game. To implement this functionality,
a new AbstractPlayer class should be created to minimize the code
replication between the existing Player class in the roster module and
the new BoxscorePlayer class in the boxscore module. Now, both of those
classes can inherit AbstractPlayer to keep common properties, such as
'yards', 'rush_attempts', and 'touchdowns' in a singular location
instead of containing two copies of a property which would then have the
potential to diverge.

Another goal was to minimize the changes to how the Player class in the
boxscore module is interacted with, and to ideally keep the parameters
and return values the same to prevent existing API breakage.

Signed-Off-By: Robert Clark <robdclark@outlook.com>