Improve integration test performance
Many of the integration tests were getting stuck while either trying to
pull from the external network (which should be avoided) or re-pulling
information that was unnecessary for a given test. This would cause some
tests to takeas long as two minutes while trying to pull all of the
desired information when they should be taking a couple seconds or less.
The new updates test all of the same code, but go about it in a smarter
way to reduce the testing time by 66%.

Signed-Off-By: Robert Clark <robdclark@outlook.com>