Fix issue pulling historical NCAAB fields
Prior to the 2009 season, sports-reference.com contained less data for
each NCAAB team. Any fields that were missing that require mathematical
operations (such as the subtraction in defensive_rebounds,
opp_defensive_rebounds, and net_rating), if at least one of the fields
is empty, the command will throw an error. To combat this behavior, we
need to pre-emptively catch any arithmetic errors and return a default
value.

Signed-Off-By: Robert Clark <robdclark@outlook.com>