Revert to previous season with invalid URL
Just prior to the start of a new season, the URL that is provided by
sportsreference will be invalid as it points to the newest season which
is not populated on www.sports-reference.com yet. In this case, the
previous season should be used instead as it should still be valid and
prevent errors from being thrown.

Signed-Off-By: Robert Clark <robdclark@outlook.com>