Drop support for Python 2.7
With Python 2.7 reaching its end-of-life date on January 1, 2020,
sportsreference should officially drop support for Python 2 to encourage
users to migrate to a newer version of Python that remain supported in
the future. Only issues for Python 2.7 that pose serious security risks
will be fixed in the future and only on a case-by-case basis.

Signed-Off-By: Robert Clark <robdclark@outlook.com>