Fix issue pulling weight of older NCAAF players
While attempting to pull player information for players from several
seasons ago (tested on 2000 and earlier), sports-reference.com doesn't
always contain the player's weight. In this case, their weight will
still be set as None, causing an AttributeError to be thrown while
attempting to replace the contents of what it expects to be a string.
If this situation occurs, the weight should stay as None for the
end-user to decide how they want to handle the omission.

Signed-Off-By: Robert Clark <robdclark@outlook.com>