Fix issue pulling NCAAF games without time
In the seasons prior to 2013, sports-reference.com does not include a
time field in the NCAAF game's information page. This causes the time
property to be set to None, which fails while attempting to parse a
DateTime between the date and time properties. This behavior can be
circumvented by only creating a DateTime of the date as opposed to the
date and time when time is None.

Signed-Off-By: Robert Clark <robdclark@outlook.com>