Fix issue pulling NFL rushing and receiving data
The name of the table containing player statistics for rushing and
receiving has been changed on sports-reference.com from
'rushing_and_receiving' to 'receiving_and_rushing'. To hopefully combat
a change again in the future, both table names should be checked. If one
is empty, it will just be ignored.

Signed-Off-By: Robert Clark <robdclark@outlook.com>