Update Houston Texan's abbreviation
The team at sports-reference.com changed the abbreviation used for the
Houston Texans in all of the NFL's pages from 'HOU' to 'HTX'. When
attempting to retrieve data from the Texans using 'HOU', an error will
be thrown indicating that there is no data to collect. To reflect this
change, all of the documentation that mentioned 'HOU' in the NFL
modules should be updated to 'HTX'.

Signed-Off-By: Robert Clark <robdclark@outlook.com>