Fix issue retrieving data from empty pages
Occasionally, such as right before the beginning of a season, a page
will return a valid response but will have no data outside of the
default HTML and CSS template. With no data present on the page,
sportsreference can't parse any information and should indicate the lack
of data and return safely.

Signed-Off-By: Robert Clark <robdclark@outlook.com>