Fix NCAAF conference class for empty pages
Occasionally, a conference page for NCAAF will be an empty document,
devoid of any information that can be parsed. When this happens, the
Teams class throws an error as it relies on getting information from the
Conferences class to determine which conference each team participates
in. Instead of attempting to pull the information from other sources, if
a conference page isn't valid, any team that participates in that
conference should have a conference attribute of None to avoid
presenting incorrect information.

Signed-Off-By: Robert Clark <robdclark@outlook.com>