Set default NCAAB time to midnight
In case an NCAAB game doesn't have a time value associated with it, the
default time should be set to midnight instead of 7:00PM as users should
have the option on how to handle missing data instead of providing false
information. Now, users can check if the `time` property is empty and
determine how they want to treat the data in such cases.

Signed-Off-By: Robert Clark <robdclark@outlook.com>