Fix identical home and away scores in NBA Boxscore
The NBA Boxscores module contains an issue where both the home and away
points will return just the away team's points. The HTML has changed on
the website in a way that multiple scores are returned instead of just a
singular score for home and away, throwing the indexing off for
selecting the home team's points. Selecting the final index for the home
team's points total works in the case that multiple instances are used
for the score.

Signed-Off-By: Robert Clark <robdclark@outlook.com>