Fix HTTP redirect issues in NCAAB module
The NCAAB module was throwing unhandled HTTP 301 status codes for all
pages, which masked 404 errors in the case of pages not being present,
and therefore allowing false positives for pulling data from the
website. The 301 codes were being thrown as the server automatically
redirects from http:// to https:// for the linked pages, and should
therefore default to https:// for the URLs to ensure proper codes are
always returned while pulling webpages.

Signed-Off-By: Robert Clark <robdclark@outlook.com>