Resolve errors while pulling empty NBA stats pages
Some players don't have a personal stats page created for them yet, but
contain a valid link on roster pages. As implemented, the Roster module
for the NBA does not check if the returned data from the page is empty,
and will attempt to parse information from an object which is None.
Instead, if a player's page is invalid, the module should fail-fast and
return None for the player instead of attempting to parse information or
handle other scenarios.

Signed-Off-By: Robert Clark <robdclark@outlook.com>