Fix issue with future NFL boxscore errors
For games that are yet to occur which are listed in an NFL team's
schedule, an error will be thrown while attempting to parse a score for
the game when none exists. To circumvent this issue, not only should the
points be checked if they are None, they should also be checked if they
are empty.

Signed-Off-By: Robert Clark <robdclark@outlook.com>