Move exhaustive tests
The naming structure used for the exhaustive tests forced them to be
processed by the testing infrastructure, dramatically increasing the
time required to both discover and run tests over the code. The tests
are designed to be run in a few minutes, but the exhaustive tests take
several hours to complete and should be run overnight instead of on
every check-in. Changing the test names prevents py.test from
identifying them as tests to execute.

Signed-Off-By: Robert Clark <robdclark@outlook.com>