Trigger actions on Pull Requests
The checks run by GitHub Actions are only triggered when a branch is
pushed to the parent (non-fork) repository, but don't run when a PR is
created. This is an issue when anyone creates a PR based on a fork they
have created, preventing any collaborators from having their code
reviewed automatically.

Signed-Off-By: Robert Clark <robdclark@outlook.com>