Add football Roster class
The Roster class can be used to retrieve detailed stats for each player
on a team's roster for the given season, ranging from goals, assists,
and time-related statistics. The roster can be queried by supplying
either the squad ID or the full team name. Every player will be an
instance of the Player class in the same module.

Signed-Off-By: Robert Clark <robdclark@outlook.com>