Fix issue pulling missing NCAAB player weight
Occasionally, some NCAAB players will not have a weight listed on their
stats page on sports-reference.com. This will return `None` for the
property in the Player class. While attempting to parse the weight from
None, an error will be thrown. Doing a simple check to see if the weight
is valid prior to parsing will resolve the issue and allow execution as
expected.

Signed-Off-By: Robert Clark <robdclark@outlook.com>