Skip NCAAB teams without stat pages
In 2010 and earlier years, some NCAAB teams, such as Bryant University,
made the transition from Division-II or lower to Division-I basketball.
This caused an issue on the website as sometimes those teams would be
displayed in the overall school stats, but they would not have their own
standalone stats page specific to their school. This would prevent the
code from parsing the team's conference and other stats, and would throw
an error while attempting to instantiate the team. Doing a simple check
to see if the team exists in the conference dataset prior to checking
instantiating the team mitigates this issue.

Signed-Off-By: Robert Clark <robdclark@outlook.com>