Convert MLB innings pitched to float
The BoxscorePlayer class for MLB currently forces the number of innings
pitched per player to be an integer, truncating any partial innings they
might have played. This makes the data incomplete and unreliable.
Allowing the number of innings pitched to be a float will now indicate
partial innings pitched, with results ending in '.1' or '.2' to indicate
one third or two thirds of an inning pitched, respectively.

Signed-Off-By: Robert Clark <robdclark@outlook.com>