Switch MLB Boxscore player teams
The home and away players in the MLB Boxscore class were being saved in
the opposite property, causing home players to show under away_players
and vice-versa. The algorithm wrongly assumed the away team matched with
odd-numbered tables, when the numbers are zero-based, and should instead
match with even-numbered tables as a result.

Signed-Off-By: Robert Clark <robdclark@outlook.com>