Fix NFL datetimes for playoff games
For NFL teams that played in playoff games which took place after the
New Year, the datetime in the schedule would be incorrect as it would
use the default year instead of the actual year the game was played.
Checking if the game was played in the first couple months of the year
and incrementing the year by 1 if so resolves this issue as games being
played in January or February of the same season are in a different
calendar year.

Signed-Off-By: Robert Clark <robdclark@outlook.com>