Fix issue pulling football schedules and rosters
If the Team class for football is unable to pull a team's page, the
document for the team will be invalid, and the '_doc' attribute will not
exist for that team. While trying to pull the schedule or roster for
that team, an error will be thrown that the '_doc' attribute does not
exist.

Simply checking if the attribute exists, and setting it to None if not
will allow the Schedule and Roster modules to retry pulling the page
again to see if it works.

Signed-Off-By: Robert Clark <robdclark@outlook.com>