Use tokenless Codecov action
The original version of Codecov used in the repository required a token
for proper authentication, which worked fine for all non-fork updates to
the repository as the GitHub action secret was saved and usable. For all
forked updates, however, the secret was not saved, and all actions would
fail as Codecov would try and push the coverage report without a token.

With the latest release of Codecov action v1.0.6, tokenless pushing is
now available which solves the problem of pushing from forked
repositories.

Signed-Off-By: Robert Clark <robdclark@outlook.com>