Fix issue with FB schedule using invalid date (#409)
Using the current datetime could run into an issue for the football
schedule tests if the date a test was run happened to be an actual date
on the team's schedule would fail to throw a ValueError. Instead, a date
that is known to not be on the team's schedule should be used to ensure
the tests can complete regardless of the date run.

Signed-Off-By: Robert Clark <robdclark@outlook.com>