Drop support for Python 3.5
With Python 3.5 reaching its end-of-life date on September 13, 2020,
sportsreference should officially drop support for Python 3.5 to
encourage users to migrate to a newer version of Python that remain
supported in the future. Only issues for Python 2.7 that pose serious
security risks will be fixed in the future and only on a case-by-case
basis.

Signed-Off-By: Robert Clark <robdclark@outlook.com>