Fix issues pulling default NBA year for 2020
Given the lengthy delays to the NBA season in 2020, attempting to pull
the NBA modules towards the end of the calendar year throws off the
logic as the default logic assumes the next season is either getting
ready to start or is already underway. In 2020, however, this has not
yet happened, and an HTTPError will be thrown while attempting to pull
the stats without specifying the year. By doing a one-off check for
2020, a handler can be made to identify whether or not the next season
has started yet. This will only be applicable for the second half of
2020, and can be removed at a later date.

Signed-Off-By: Robert Clark <robdclark@outlook.com>