Fix NBA home team score in Boxscore
The Boxscore class was producing the incorrect score for the home team,
and instead was pulling the home team's second half points total. By
reading the score directly from the header, the chances of running into
a parsing issue are much lower in case the layout changes again on the
website.

Signed-Off-By: Robert Clark <robdclark@outlook.com>