Fix NBA Boxscores using the wrong home values
The Boxscores pages on sports-reference.com were changed to include
scores by halves and quarters, which retrieved the incorrect values for
the home team as it was instead pulling the first-quarter stats for the
away team. By using a new index to retrieve values from, the proper
stats can be displayed for the home team.

Signed-Off-By: Robert Clark <robdclark@outlook.com>