Fix NFL players wrongly being listed as away
Occasionally, some NFL Boxscores would incorrectly list the home team
players as away. This is caused by the home and away teams using
different abbreviations in their URIs than the abbreviations listed in
the tables, which would fail the branching logic to select the players
for a specific team.

By checking the abbreviations listed on the actual page outside of the
URIs, the players can be placed properly into one team or the other.

Signed-Off-By: Robert Clark <robdclark@outlook.com>