Update default values for NBA players
In cases where no information is present for an NBA player, the default
value of None should be used for all attributes. At present, the code
was unable to initialize many default common variables for players,
which resulted in an error while trying to pull data from the Roster
class.

Signed-Off-By: Robert Clark <robdclark@outlook.com>