Handle redirects to 404 errors
If the code throws a 301 redirect error which leads to a 404 error, it
will be treated as the request returning a valid result, when that
shouldn't be the case. By pulling the page on a status code of 301, the
final value is returned for the status_code, which in the case of a 404
error, will return the intended value.

Signed-Off-By: Robert Clark <robdclark@outlook.com>