Fix empty NFL Rosters
Currently, when trying to get the roster for a given NFL team, the
result is an empty list. The reason for this is because the html element
that is being used to search within the roster html page is incorrect.
When changing that element to an element that exists on the page, we get
the correct results for an NFL roster.

Signed-Off-By: Alec Keller <alec.keller.17@gmail.com>